feat(auth): 实现数据库认证方式
- 新增 DbAuth 结构体和相关实现，支持通过数据库进行用户认证
- 修改认证中间件，支持 JWT 和数据库两种认证方式
- 更新应用配置，添加认证方法配置项
- 优化 Redis 缓存操作，支持 JSON 数据缓存

+++ crates/daoyi-entities-system/src/entity/system_oauth2_access_token.rs
@@ -10,7 +10,7 @@ pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
     pub id: i64,
     pub user_id: i64,
-    pub user_type: i16,
+    pub user_type: i32,
     pub user_info: String,
     pub access_token: String,
     pub refresh_token: String,
